Add flat_tile_transform and globe_tile_transform
